# webapi/gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"Students API","description":"API for student information management","version":"1.0"},"servers":[{"url":"http://localhost:8080"}],"paths":{"/students":{"get":{"tags":["students"],"summary":"get students students","description":"学籍番号で昇順にソートされた全ての学生を取得する。","operationId":"students#get students","responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Students"},"example":{"students":[{"address":"名古屋市中区三の丸三丁目1番2号","date_of_birth":"2022-04-01T13:30:00+09:00","expiration_date":"2027-03-31T00:00:00+09:00","id":1,"name":"鈴木太郎","ruby":"スズキタロウ","student_number":12345},{"address":"名古屋市中区三の丸三丁目1番2号","date_of_birth":"2022-04-01T13:30:00+09:00","expiration_date":"2027-03-31T00:00:00+09:00","id":1,"name":"鈴木太郎","ruby":"スズキタロウ","student_number":12345},{"address":"名古屋市中区三の丸三丁目1番2号","date_of_birth":"2022-04-01T13:30:00+09:00","expiration_date":"2027-03-31T00:00:00+09:00","id":1,"name":"鈴木太郎","ruby":"スズキタロウ","student_number":12345},{"address":"名古屋市中区三の丸三丁目1番2号","date_of_birth":"2022-04-01T13:30:00+09:00","expiration_date":"2027-03-31T00:00:00+09:00","id":1,"name":"鈴木太郎","ruby":"スズキタロウ","student_number":12345}]}}}},"500":{"description":"internal_error: Internal Server Error response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomError"}}}}}},"post":{"tags":["students"],"summary":"create student students","description":"学生を登録する。","operationId":"students#create student","responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"},"example":{"address":"名古屋市中区三の丸三丁目1番2号","date_of_birth":"2022-04-01T13:30:00+09:00","expiration_date":"2027-03-31T00:00:00+09:00","id":1,"name":"鈴木太郎","ruby":"スズキタロウ","student_number":12345}}}},"400":{"description":"bad_request: Bad Request response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomError"}}}},"500":{"description":"internal_error: Internal Server Error response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomError"}}}}}}},"/students/{id}":{"get":{"tags":["students"],"summary":"get student students","description":"id から学生を取得する。","operationId":"students#get student","parameters":[{"name":"id","in":"path","description":"Student's unique ID","required":true,"schema":{"type":"integer","description":"Student's unique ID","example":2311868056971056429,"format":"int64"},"example":7361208326735898235}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"},"example":{"address":"名古屋市中区三の丸三丁目1番2号","date_of_birth":"2022-04-01T13:30:00+09:00","expiration_date":"2027-03-31T00:00:00+09:00","id":1,"name":"鈴木太郎","ruby":"スズキタロウ","student_number":12345}}}},"404":{"description":"not_found: Not Found response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomError"}}}},"500":{"description":"internal_error: Internal Server Error response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomError"}}}}}}}},"components":{"schemas":{"CustomError":{"type":"object","properties":{"message":{"type":"string","description":"Message of error","example":"This is an error message."},"name":{"type":"string","description":"Name of error","example":"internal_error"}},"example":{"message":"This is an error message.","name":"internal_error"},"required":["name","message"]},"Student":{"type":"object","properties":{"address":{"type":"string","description":"学生の住所","example":"名古屋市中区三の丸三丁目1番2号"},"date_of_birth":{"type":"string","description":"学生の生年月日 (RFC3339)","example":"2022-04-01T13:30:00+09:00","format":"date-time"},"expiration_date":{"type":"string","description":"学生証の有効期間 (RFC3339)","example":"2027-03-31T00:00:00+09:00","format":"date-time"},"id":{"type":"integer","description":"学生を一意に表す ID","example":1,"format":"int64"},"name":{"type":"string","description":"学生の氏名","example":"鈴木太郎"},"ruby":{"type":"string","description":"学生の氏名のフリガナ","example":"スズキタロウ"},"student_number":{"type":"integer","description":"学生の学籍番号","example":12345,"format":"int64"}},"example":{"address":"名古屋市中区三の丸三丁目1番2号","date_of_birth":"2022-04-01T13:30:00+09:00","expiration_date":"2027-03-31T00:00:00+09:00","id":1,"name":"鈴木太郎","ruby":"スズキタロウ","student_number":12345},"required":["id","name","ruby","student_number","date_of_birth","address","expiration_date"]},"Student2":{"type":"object","properties":{"address":{"type":"string","description":"学生の住所","example":"名古屋市中区三の丸三丁目1番2号"},"date_of_birth":{"type":"string","description":"学生の生年月日 (RFC3339)","example":"2022-04-01T13:30:00+09:00","format":"date-time"},"expiration_date":{"type":"string","description":"学生証の有効期間 (RFC3339)","example":"2027-03-31T00:00:00+09:00","format":"date-time"},"id":{"type":"integer","description":"学生を一意に表す ID","example":1,"format":"int64"},"name":{"type":"string","description":"学生の氏名","example":"鈴木太郎"},"ruby":{"type":"string","description":"学生の氏名のフリガナ","example":"スズキタロウ"},"student_number":{"type":"integer","description":"学生の学籍番号","example":12345,"format":"int64"}},"description":"One student","example":{"address":"名古屋市中区三の丸三丁目1番2号","date_of_birth":"2022-04-01T13:30:00+09:00","expiration_date":"2027-03-31T00:00:00+09:00","id":1,"name":"鈴木太郎","ruby":"スズキタロウ","student_number":12345},"required":["id","name","ruby","student_number","date_of_birth","address","expiration_date"]},"Students":{"type":"object","properties":{"students":{"type":"array","items":{"$ref":"#/components/schemas/Student2"},"example":[{"address":"名古屋市中区三の丸三丁目1番2号","date_of_birth":"2022-04-01T13:30:00+09:00","expiration_date":"2027-03-31T00:00:00+09:00","id":1,"name":"鈴木太郎","ruby":"スズキタロウ","student_number":12345},{"address":"名古屋市中区三の丸三丁目1番2号","date_of_birth":"2022-04-01T13:30:00+09:00","expiration_date":"2027-03-31T00:00:00+09:00","id":1,"name":"鈴木太郎","ruby":"スズキタロウ","student_number":12345},{"address":"名古屋市中区三の丸三丁目1番2号","date_of_birth":"2022-04-01T13:30:00+09:00","expiration_date":"2027-03-31T00:00:00+09:00","id":1,"name":"鈴木太郎","ruby":"スズキタロウ","student_number":12345},{"address":"名古屋市中区三の丸三丁目1番2号","date_of_birth":"2022-04-01T13:30:00+09:00","expiration_date":"2027-03-31T00:00:00+09:00","id":1,"name":"鈴木太郎","ruby":"スズキタロウ","student_number":12345}]}},"example":{"students":[{"address":"名古屋市中区三の丸三丁目1番2号","date_of_birth":"2022-04-01T13:30:00+09:00","expiration_date":"2027-03-31T00:00:00+09:00","id":1,"name":"鈴木太郎","ruby":"スズキタロウ","student_number":12345},{"address":"名古屋市中区三の丸三丁目1番2号","date_of_birth":"2022-04-01T13:30:00+09:00","expiration_date":"2027-03-31T00:00:00+09:00","id":1,"name":"鈴木太郎","ruby":"スズキタロウ","student_number":12345},{"address":"名古屋市中区三の丸三丁目1番2号","date_of_birth":"2022-04-01T13:30:00+09:00","expiration_date":"2027-03-31T00:00:00+09:00","id":1,"name":"鈴木太郎","ruby":"スズキタロウ","student_number":12345}]},"required":["students"]}}},"tags":[{"name":"students"}]}